requirements.txt
uvicorn==0.17.6
fastapi==0.78.0
python-dotenv

cassandra-driver==3.25.0